+++ CHANGELOG.rst
@@ -15,6 +15,13 @@ are used for versioning (schema follows below):
   0.3.4 to 0.4).
 - All backwards incompatible changes are mentioned in this document.
 
+0.1.4
+-----
+2023-09-23
+
+- Set minimal required version of faker-file to 0.17.8.
+- Add selects for image, pdf and mp3 generator selection.
+
 0.1.3
 -----
 2023-07-28